Add graph resizing to answer page
